# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KATALA - Katalog & Reservasi Wisata Lampung API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reservations:
    get:
      operationId: listMyReservations
      tags: [reservations]
      summary: List current user's reservations
      responses:
        "200":
          description: User reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          description: Quota exceeded or invalid
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get reservation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"

  # ─── TICKETS ─────────────────────────────────────────────
...
  /payments/create:
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create Midtrans payment transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCreateInput"
      responses:
        "201":
          description: Payment snap token created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /payments/webhook:
    post:
      operationId: paymentWebhook
      tags: [payments]
      summary: Midtrans payment notification webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookPayload"
      responses:
        "200":
          description: Webhook received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /payments/{reservationId}/status:
    get:
      operationId: getPaymentStatus
      tags: [payments]
      summary: Get payment status for a reservation
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # ─── ADMIN ───────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ─── AUTH ─────────────────────────────────────────────
...
    Destination:
      type: object
      required: [id, name, description, location, latitude, longitude, price, dailyQuota, category, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        location:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        price:
          type: number
        dailyQuota:
          type: integer
        category:
          type: string
        imageUrl:
          type: string
        rating:
          type: ["number", "null"]
        featured:
          type: boolean
        createdAt:
          type: string
...
    Reservation:
      type: object
      required: [id, userId, destinationId, visitDate, quantity, totalPrice, status]
      properties:
        id:
          type: integer
        userId:
          type: integer
        destinationId:
          type: integer
        visitDate:
          type: string
        quantity:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, paid, cancelled, expired]
        destination:
          $ref: "#/components/schemas/Destination"
        createdAt:
          type: string
    ReservationInput:
      type: object
      required: [destinationId, visitDate, quantity]
      properties:
        destinationId:
          type: integer
        visitDate:
          type: string
        quantity:
          type: integer

    # ─── TICKETS ───────────────────────────────────────────
...
    PaymentCreateInput:
      type: object
      required: [reservationId]
      properties:
        reservationId:
          type: integer
    PaymentResponse:
      type: object
      required: [snapToken, orderId]
      properties:
        snapToken:
          type: string
        orderId:
          type: string
        redirectUrl:
          type: ["string", "null"]
    Transaction:
      type: object
      required: [id, reservationId, orderId, amount, status]
      properties:
        id:
          type: integer
        reservationId:
          type: integer
        orderId:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, failed, expired, refunded]
        paymentType:
          type: ["string", "null"]
        midtransTransactionId:
          type: ["string", "null"]
        reservation:
          $ref: "#/components/schemas/Reservation"
        createdAt:
          type: string
        updatedAt:
          type: string
    WebhookPayload:
      type: object
      required: [order_id, transaction_status, fraud_status]
      properties:
        order_id:
          type: string
        transaction_status:
          type: string
        fraud_status:
          type: string
        payment_type:
          type: ["string", "null"]
        transaction_id:
          type: ["string", "null"]
        gross_amount:
          type: ["string", "null"]

    # ─── ADMIN ─────────────────────────────────────────────
...