# lib/api-zod/src/generated/types/paymentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KATALA - Katalog & Reservasi Wisata Lampung API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentResponse {
  snapToken: string;
  orderId: string;
  /** @nullable */
  redirectUrl?: string | null;
}